requirements.txt
pytest==8.4.1
Faker==37.4.0
Appium-Python-Client==2.0.1
selenium==4.34.0
